move types file from root to types folder
